Migrate from JavaScript to TypeScript
Since this project unfortunately lacks tests for the time being, I tried
to leave as much of the code as possible unchanged, doing my best to
restrict my changes to type annotations and assertions.
As a result, some of the code may not be the most idiomatic TypeScript.
For example, the code may have a lot of awkward type casts and non-null
assertions, and files that would ideally take advantage of TypeScript
niceties like enums and discriminated unions don't do so.
We can worry about making the code more idiomatic in a later commit.